disputes flow
## what
updated dispute base. Now replacing selections with a custom input selections. Even if already having one or two selections made.

# how
added additional logic to check if the selections is >= to new max length and also if the reason selected  allowsInput, if so them remove all previous selection, updating the selections from the carousel as well as the current selections